Blackberry Z10 Pokedex App
This is a poked like app developed with some help from one of my
professors for one of my classes. It has data about all 718 pokemon and
supports, to some extent, 9 different languages. It was written using
the QNX Momentics IDE and can only run on a Blackberry Z10 device or a
virtual machine running its OS. It also has an awkward 7 second load
time at start up and when “All Types” is selected. This is from the
nature of the class project requiring the use of multiple .cvs files
acquired form GitHub.